refactor: tooltip 이름 변경 및 선언 합치기
