+++ src/views/baseball/ui/Heatmap/index.css.ts
@@ -49,25 +49,3 @@ export const week = style({
   gap: '3px',
   flexShrink: 0,
 });
-
-/**
- * ======== FLOATING TOOLTIP ========
- */
-
-export const floatingTooltip = style({
-  position: 'fixed',
-  transform: 'translate(10px, -50%)',
-  zIndex: 1000,
-
-  fontFamily: vars.font.sans,
-  fontSize: '11px',
-  lineHeight: 1.6,
-  whiteSpace: 'nowrap',
-
-  color: '#fff',
-  backgroundColor: vars.color.text.strong,
-  padding: '8px 12px',
-  borderRadius: '6px',
-
-  pointerEvents: 'none',
-});
